Change the dependency type of `xposed-javax` to `implementation`
This should be the correct thing to do, as `xposed-javax:1.2.0` is not something provided by the runtime environment, but rather a dependency that we have to include in our build artifact.

In order to reliably include `xposed-javax:1.2.0` in the build output, one should rely on Gradle's "transitive dependencies" feature when declaring this project as a dependency.